requirements.txt
google-generativeai
bs4
faiss-cpu
PyPDF2
IPython
gtts
SpeechRecognition
streamlit
tiktoken
python-dotenv